% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPeople.R
\name{GetUnlabeledID}
\alias{GetUnlabeledID}
\title{Get an unlabeled entity}
\usage{
GetUnlabeledID(path_data = NULL, labelee_data = NULL, path_labels)
}
\arguments{
\item{path_data}{path to data we want to label}

\item{path_labels}{path to file where we track previously labeled data}
}
\value{
ourID the ID of an unlabeled entity
}
\description{
Find someone or something that has not been previously labeled
}
\examples{

}
