% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadClean.R
\name{LoadCleanRaw_fread}
\alias{LoadCleanRaw_fread}
\title{Consistently load raw Irish census data}
\usage{
LoadCleanRaw_fread(path, file_delim = " ", assignID = NULL, preProcess = FALSE)
}
\arguments{
\item{path}{path to data}

\item{file_delim}{single character used to separate fields within a record; default is " " for .txt files}

\item{assignID}{string of variable names for creating a unique identifier}

\item{preProcess}{has the data been preprocessed?}
}
\value{
data in tibble format
}
\description{
Consistently load and standardize raw Irish census data
}
\examples{
LoadCleanRaw("~/GoogleDrive/irelandData/ticknock_kayla.csv", ",", preProcess = TRUE)
LoadCleanRaw("~/GoogleDrive/irelandData/census_ireland_1901/Carlow/Ticknock.txt")
LoadCleanRaw("~/GoogleDrive/irelandData/census_ireland_1901/Carlow/Ticknock.txt",
 assignID = c("County", "DED", "Year"))
}
